+++ solution.hpp
@@ -6,16 +6,16 @@
 class Solution
 {
 	private:
-		std::vector<int> parcours;
+		std::vector<double> parcours;
 		int N;
 		int score;
 		
 	public:
 		Solution();
 		void afficher() const;
-		Solution(const std::vector<int>&, int, const std::vector<std::vector<int> >&);
+		Solution(const std::vector<double>&, int, const std::vector<std::vector<double> >&);
 		bool operator==(const Solution&) const;
-		bool operator==(const std::vector<int>&) const;
+		bool operator==(const std::vector<double>&) const;
 };
 
 #endif
